+++ NEWS.md
@@ -8,3 +8,13 @@
         * (Resizing is not ~yet~ supported (thankfully :/))
     * And so that's what I did. It works! Yay! (for now...)
     * The Win32 API is designed by a monkey on fent and you won't convince me otherwise
+
+* Made the return value (errors) of `p_window_swap_buffers` not retarded
+    * Since a `NULL` return value might actually sometimes indicate success
+        (e.g. when `P_WINDOW_ACCELERATION_OPENGL` is used),
+        added a special value - `P_WINDOW_SWAP_BUFFERS_FAIL` to indicate failure of the request.
+    * Implemented this change to the API in `platform/windows`
+    * Also fixed the documentation to make it clear that `p_window_swap_buffers` only "requests" the
+        presentation and doesn't block until any results of the actual page flip are available.
+    * Fixed a bunch of typos in `platform/window.h` & generally improved the documentation of `p_window_swap_buffers`
+    * Fixed the `timestamp_delta` macro in `platform/ptime.h`